refactor: moved components inside folders
