+++ components/tools/whatsapp-message-link-generator/WhatsappLinksList.tsx
@@ -1,14 +1,14 @@
 //? Memoize the array of JSX.Elements to avoid spamming calculations
 import { useMemo } from "preact/hooks";
 //? Types for typecasting
-import type { WhatsappLinkData } from "../../types/component-properties/tools/whatsapp-link-generator/whatsapp-link-data.ts";
+import type { WhatsappLinkData } from "../../../types/component-properties/tools/whatsapp-link-generator/whatsapp-link-data.ts";
 //? Create gradient links to all numbers provided
-import { GradientLink } from "../UI/GradientLink.tsx";
+import { GradientLink } from "../../UI/GradientLink.tsx";
 //? Create a greek list of contents
-import { GreekList } from "../UI/GreekList.tsx";
+import { GreekList } from "../../UI/GreekList.tsx";
 //? Whatsapp message Variable slots
-import { WHATSAPP_MESSAGE_VARIABLE_PLACEHOLDER } from "../../data/tools/whatsapp-link-generator/whatsapp-message-variable-placeholder.ts";
-import { WHATSAPP_MESSAGE_MISSING_VARIABLE_MESSAGE } from "../../data/tools/whatsapp-link-generator/whatsapp-message-missing-variable.ts";
+import { WHATSAPP_MESSAGE_VARIABLE_PLACEHOLDER } from "../../../data/tools/whatsapp-link-generator/whatsapp-message-variable-placeholder.ts";
+import { WHATSAPP_MESSAGE_MISSING_VARIABLE_MESSAGE } from "../../../data/tools/whatsapp-link-generator/whatsapp-message-missing-variable.ts";
 
 //? Dynamically display a list of links to messaging one or more numbers on Whatsapp
 export function WhatsappLinksList(
