+++ islands/tools/WhatsappLinkGenerator.tsx
@@ -3,17 +3,17 @@ import { useEffect, useRef, useState } from "preact/hooks";
 //? Styled components
 import { StyledInput } from "../../components/UI/StyledInput.tsx";
 import { StyledButton } from "../../components/UI/StyledButton.tsx";
-import { CountryPhoneCodeSelect } from "../../components/tools/CountryPhoneCodeSelect.tsx";
+import { CountryPhoneCodeSelect } from "../../components/UI/CountryPhoneCodeSelect.tsx";
 //? Types for typecasting
 import { validationStatus } from "../../types/forms/validationStatus.ts";
 import type { WhatsappLinkData } from "../../types/component-properties/tools/whatsapp-link-generator/whatsapp-link-data.ts";
-import { WhatsappMessagePreview } from "../../components/tools/WhatsappMessagePreview.tsx";
+import { WhatsappMessagePreview } from "../../components/tools/whatsapp-message-link-generator/WhatsappMessagePreview.tsx";
 
 //? Validation functions
 import { patternValidation } from "../../services/form-validation/patternValidation.ts";
 import { validateAreaCode } from "../../services/form-validation/validateAreaCode.ts";
 import { validatePhoneNumber } from "../../services/form-validation/validatePhoneNumber.ts";
-import { WhatsappLinksList } from "../../components/tools/WhatsappLinksList.tsx";
+import { WhatsappLinksList } from "../../components/tools/whatsapp-message-link-generator/WhatsappLinksList.tsx";
 import { StyledTextArea } from "../../components/UI/StyledTextArea.tsx";
 
 //? Defined variable to replace in message body
